// src/features/network-modifications/linesAttachToSplittingLines/linesAttachToSplittingLines.type.ts
/*
 * Copyright (c) 2026, RTE (http://www.rte-france.com)
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 * SPDX-License-Identifier: MPL-2.0
 */

import { UUID } from 'node:crypto';
import { ModificationType } from '../../../utils';

export interface LinesAttachToSplittingLinesDto {
    type: ModificationType.LINES_ATTACH_TO_SPLIT_LINES;
    uuid?: UUID;
    lineToAttachTo1Id: string;
    lineToAttachTo2Id: string;
    attachedLineId: string;
    voltageLevelId: string | null;
    bbsBusId: string | null;
    replacingLine1Id: string;
    replacingLine1Name: string | null;
    replacingLine2Id: string;
    replacingLine2Name: string | null;
}